feat: keep alternate buffer when previewing buffer
note:
when previewing buffer, because bufstop use `b` command, it will
change the alternate and current file which in most cases is not
ideal.

this changes keep the alternate buffer if we're not previewing the
alternate buffer.

the downside of this approach is that, if we use `j` and `k` for
previewing the buffer before/after alternate buffer, it will still
change the alternate buffer.

still not sure how to fix that, at the moment.